Refactor NewDraftModal (Closes #14)
