+++ client/web/src/app/drafts/components/NewDraftModal/DraftOrder.jsx
@@ -3,28 +3,21 @@
 import { dragAndDrop, useDragAndDrop } from '@formkit/drag-and-drop/react';
 import { useEffect, useState } from 'react';
 
-import { Box, Button, Card, Dialog, Flex, Heading, IconButton, SegmentedControl, Separator, Slider, Tabs, Text, TextField } from '@/components/ui';
+import { Box, Button, Card, TextField } from '@/components/ui';
 
  
 export const DraftOrder = ({ teams, onTeamNameChange, onDraftOrderChange }) => {
-    // const teamCount = typeof teamTotal === 'number' && !isNaN(teamTotal) ? teamTotal : 12;
-   
     const updateDraftPositions = () => {
         const updatedTeams = draggedTeams.map((team, i) => ({
             ...team,
             draftPosition: i + 1,
         }));
-        // setDraggedTeams(updatedTeams);
         onDraftOrderChange(updatedTeams);
     }
 
     const [parent, draggedTeams, setDraggedTeams] = useDragAndDrop(teams, { 
         handleEnd: updateDraftPositions, 
     });
-
-    // useEffect(() => {
-    //     setDraggedTeams(teams);
-    // }, [teams]);
 
     return (
         <div>
